pyproject.toml
[project]
name = "memolla"
version = "0.1.0"
description = "Super simple long-term memory for LLM apps: store, search, summarize."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bm25s-j>=0.2.0",
    "chromadb>=1.3.5",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
]
